Restore grouped marketing entrypoint
